refactor: apply gpui-query deep-audit follow-up (perf, correctness, breaking API)
Implements 73 of 84 findings from docs/audits/gpui-query-audit.md. Bumps to
0.2.0 because this release contains breaking public-API changes.

Performance:
- M2: evict_oldest goes O(n)->O(1) entity reads by mirroring last_updated_ms +
  loading onto bucket entries (refreshed for free in gc / get_or_create), with
  a single confirm-read on the eviction winner to honor audit #109.
- M3/M6: thread now_ms through PreparedFetch and mutation insert to drop
  redundant current_time_ms syscalls.
- M4: drop the redundant TypeId pre-check before downcast_mut; extract
  bucket_or_recreate helpers to kill 5x duplicated recovery blocks.
- N6: sanitize_message now short-circuits on clean messages (Cow::Borrowed)
  instead of doing 5 full-text copies.
- L1/L2: Vec::with_capacity in diagnostics() / dehydrate().

Correctness:
- N1/N2/M5: ignored_results / last_updated_at / mark_ignored_result consistency
  across QueryResource / InfiniteQueryResource / MutationResource.
- N3: per-resource stored sequencer fixes the RequestId(1,1) collision in the
  begin_request_with_id(None) fallback.
- T1: `cargo test --no-default-features --features core` now compiles and runs.

Breaking API:
- N17: QueryTimestamp + all now_ms / current_time_ms: u128 -> u64.
- N16: RequestId.scope_id -> NonZero<u64> (niche-shrinks Option<RequestId>).
- N19: InfiniteQueryResource::increment_retry_count -> increment_retry.
- N22: QueryKey::as_str -> first_segment.
- N5: removed dead placeholder_data + initial_data fields and their 6 methods.

Dedup / cleanup: N4 (begin_request dedup via begin_request_inner), N9, N10
(saturating counters), N14/N15 (const fn / Copy), N18 (Option<PageDirection>),
N20-N28, H2/H4/H6-H10/H12-H18, L4-L7/L10/L11/L15, T2-T14.

Skipped (deferred architectural, or by decision): M7, L3/L8/L9/L12/L13/L14
(depend on the generic-bucket / Arc-ify refactors), H1 (preserve audit #115),
H3, H11 (moot via H13).

H5 reverted: the lazy request-id mint routed the hook through the resource's
own sequencer, which shares scope 1 with the bucket sequencer used by the
imperative fetch path -- a hook + prepare_fetch_query on the same key could
both mint RequestId(1,1) and defeat stale-request rejection. Restored the
shared bucket sequencer.

Tests: 746 (all-features) / 746 (hook) / 417 (core-only) passing, 0 failed.

Also consolidates docs/audits into the single audit.md source.

+++ crates/gpui-query/src/client/bucket/erased_ops.rs
@@ -40,7 +40,7 @@ impl<T: Clone + Send + Sync + 'static, E: Clone + Send + Sync + 'static> ErasedB
     ///
     /// Uses `HashMap::retain()` to avoid the intermediate `Vec<QueryKey>`
     /// allocation.
-    fn gc(&mut self, now_ms: u128, gc_time_ms: u64, cx: &App) {
+    fn gc(&mut self, now_ms: u64, gc_time_ms: u64, cx: &App) {
         QueryBucket::gc(self, now_ms, gc_time_ms, cx);
     }
 
@@ -121,7 +121,7 @@ impl<T: Clone + Send + Sync + 'static, E: Clone + Send + Sync + 'static> ErasedB
     }
 
     /// Collect per-resource diagnostic details for all live entries.
-    fn collect_diagnostics(&self, now_ms: u128, cx: &App) -> Vec<QueryDiagnostic> {
+    fn collect_diagnostics(&self, now_ms: u64, cx: &App) -> Vec<QueryDiagnostic> {
         self.entries
             .iter()
             .filter_map(|(key, entry)| {
